@@ -19,13 +19,13 @@
     "@parcel/packager-ts": "2.12.0",
     "@parcel/transformer-inline-string": "2.12.0",
     "@parcel/transformer-typescript-types": "2.12.0",
-    "@size-limit/preset-big-lib": "11.1.4",
-    "prettier": "2.8.1",
-    "size-limit": "7.0.8",
+    "@size-limit/preset-big-lib": "11.2.0",
+    "prettier": "3.5.3",
+    "size-limit": "11.2.0",
     "standard-version": "9.5.0",
-    "syncpack": "12.4.0",
-    "turbo": "2.1.1",
-    "typescript": "5.1.6"
+    "syncpack": "13.0.4",
+    "turbo": "2.5.4",
+    "typescript": "5.8.3"
   },
   "scripts": {
     "build": "turbo run build",
@@ -42,10 +42,10 @@
     "test": "turbo run test"
   },
   "volta": {
-    "node": "20.17.0",
-    "pnpm": "9.9.0"
+    "node": "22.16.0",
+    "pnpm": "10.12.1"
   },
-  "packageManager": "pnpm@9.7.0",
+  "packageManager": "pnpm@10.12.1",
   "workspaces": [
     "packages/*"
   ]
