@@ -49,49 +49,52 @@
     "zod": "3.25.67"
   },
   "devDependencies": {
-    "@ladle/react": "4.1.0",
+    "@eslint/compat": "1.3.0",
+    "@eslint/eslintrc": "3.3.1",
+    "@eslint/js": "9.29.0",
+    "@ladle/react": "5.0.3",
     "@tailwindcss/aspect-ratio": "0.4.2",
     "@tailwindcss/forms": "0.5.7",
     "@tailwindcss/typography": "0.5.13",
-    "@types/node": "^20.14.14",
+    "@types/node": "^22.15.31",
     "@types/react": "^19.1.8",
     "@types/react-dom": "^19.1.6",
-    "@typescript-eslint/eslint-plugin": "7.15.0",
-    "@typescript-eslint/parser": "7.15.0",
-    "@vitejs/plugin-react-swc": "3.7.0",
-    "autoprefixer": "10.4.20",
-    "eslint": "8.57.0",
-    "eslint-plugin-react-hooks": "4.6.2",
-    "eslint-plugin-react-refresh": "0.4.9",
+    "@typescript-eslint/eslint-plugin": "8.34.1",
+    "@typescript-eslint/parser": "8.34.1",
+    "@vitejs/plugin-react-swc": "3.10.2",
+    "autoprefixer": "10.4.21",
+    "eslint": "9.29.0",
+    "eslint-plugin-react-hooks": "5.2.0",
+    "eslint-plugin-react-refresh": "0.4.20",
     "npm-run-all": "4.1.5",
     "postcss": "8.4.41",
-    "prettier": "3.3.3",
-    "prettier-plugin-organize-imports": "4.0.0",
-    "prettier-plugin-tailwindcss": "0.6.5",
-    "tsx": "4.16.5",
-    "typescript": "5.4.2",
-    "vite": "5.4.0",
-    "vite-plugin-dts": "4.0.1"
+    "prettier": "3.5.3",
+    "prettier-plugin-organize-imports": "4.1.0",
+    "prettier-plugin-tailwindcss": "0.6.13",
+    "tsx": "4.20.3",
+    "typescript": "5.8.3",
+    "vite": "6.3.5",
+    "vite-plugin-dts": "4.5.4"
   },
   "peerDependencies": {
-    "@fortawesome/fontawesome-svg-core": "^6.6.0",
-    "@fortawesome/free-regular-svg-icons": "^6.6.0",
-    "@fortawesome/free-solid-svg-icons": "^6.6.0",
+    "@fortawesome/fontawesome-svg-core": "^6.7.2",
+    "@fortawesome/free-regular-svg-icons": "^6.7.2",
+    "@fortawesome/free-solid-svg-icons": "^6.7.2",
     "@fortawesome/react-fontawesome": "^0.2.2",
-    "class-variance-authority": "^0.7.0",
+    "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
-    "dayjs": "^1.11.12",
+    "dayjs": "^1.11.13",
     "formik": "^2.4.6",
-    "lucide-react": "^0.424.0",
+    "lucide-react": "^0.522.0",
     "postcss": "^8.4.41",
     "postcss-import": "^16.1.0",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
-    "tailwind-merge": "^2.4.0",
+    "tailwind-merge": "^2.6.0",
     "tailwindcss": "^3.4.7",
     "tailwindcss-animate": "^1.0.7",
     "tailwindcss-radix": "^3.0.0",
-    "yup": "^1.0.0"
+    "yup": "^1.6.1"
   },
   "scripts": {
     "build": "run-s build:tsc build:vite build:copy",
@@ -103,10 +106,11 @@
     "format": "prettier --write .",
     "format:check": "prettier --check .",
     "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
-    "preview": "LADLE=true ladle preview"
+    "preview": "LADLE=true ladle preview",
+    "check": "tsc --noEmit"
   },
   "engines": {
-    "node": "=20"
+    "node": "=22"
   },
   "volta": {
     "extends": "../../package.json"
