@@ -13,7 +13,7 @@
   },
   "devDependencies": {
     "@parcel/plugin": "2.12.0",
-    "@parcel/transformer-inline-string": "2.12.0",
+    "@parcel/transformer-inline-string": "2.15.4",
     "@types/node": "^20.0.0",
     "@types/preact-custom-element": "^4.0.1",
     "@uzh-bf/parcel-config-design-system": "workspace:*",
@@ -23,7 +23,7 @@
     "postcss": "8.4.23",
     "rimraf": "5.0.0",
     "tailwindcss": "3.3.2",
-    "typescript": "5.1.6"
+    "typescript": "5.8.3"
   },
   "scripts": {
     "build": "run-s build:purge build:parcel",
