@@ -11,9 +11,9 @@
     "tailwind-merge": "1.12.0"
   },
   "devDependencies": {
-    "@parcel/packager-ts": "2.12.0",
+    "@parcel/packager-ts": "2.15.4",
     "@parcel/plugin": "2.12.0",
-    "@parcel/transformer-typescript-types": "2.12.0",
+    "@parcel/transformer-typescript-types": "2.15.4",
     "@types/node": "^20.0.0",
     "@types/react": "^18.2.0",
     "@types/react-dom": "^18.2.1",
@@ -23,7 +23,7 @@
     "postcss": "8.4.23",
     "rimraf": "5.0.0",
     "tailwindcss": "3.3.2",
-    "typescript": "5.1.6"
+    "typescript": "5.8.3"
   },
   "scripts": {
     "build": "run-s build:purge build:parcel build:tailwind",
