@@ -1,4 +1,4 @@
-module.exports = {
+export default {
   semi: false,
   singleQuote: true,
   trailingComma: 'es5',
