@@ -5,6 +5,4 @@ import React from 'react'
 import 'tailwindcss/src/css/preflight.css'
 import 'tailwindcss/tailwind.css'
 
-export const Provider: GlobalProvider = ({ children, globalState }) => (
-  <>{children}</>
-)
+export const Provider: GlobalProvider = ({ children }) => <>{children}</>
